fix Float64 latexstrings

+++ src/basic_recipes/error_and_rangebars.jl
@@ -252,7 +252,7 @@ function plot_to_screen(plot, points::AbstractVector)
 
     return map(points) do p
         transformed = apply_transform(transform_func(plot), p, space)
-        p4d = spvm * to_ndim(Point4f, to_ndim(Point3f, transformed, 0), 1)
+        p4d = spvm * to_ndim(Point4d, to_ndim(Point3d, transformed, 0), 1)
         return Point2f(p4d) / p4d[4]
     end
 end
@@ -263,7 +263,7 @@ function plot_to_screen(plot, p::VecTypes)
     spvm = clip_to_space(cam, :pixel) * space_to_clip(cam, space) *
         f32_convert_matrix(plot, space) * transformationmatrix(plot)[]
     transformed = apply_transform(transform_func(plot), p, space)
-    p4d = spvm * to_ndim(Point4f, to_ndim(Point3f, transformed, 0), 1)
+    p4d = spvm * to_ndim(Point4d, to_ndim(Point3d, transformed, 0), 1)
     return Point2f(p4d) / p4d[4]
 end
 
@@ -275,8 +275,8 @@ function screen_to_plot(plot, points::AbstractVector)
     itf = inverse_transform(transform_func(plot))
 
     return map(points) do p
-        pre_transform = mvps * to_ndim(Vec4f, to_ndim(Vec3f, p, 0.0), 1.0)
-        p3 = Point3f(pre_transform) / pre_transform[4]
+        pre_transform = mvps * to_ndim(Vec4d, to_ndim(Vec3d, p, 0.0), 1.0)
+        p3 = Point3d(pre_transform) / pre_transform[4]
         return apply_transform(itf, p3, space)
     end
 end
@@ -286,8 +286,8 @@ function screen_to_plot(plot, p::VecTypes)
     space = to_value(get(plot, :space, :data))
     mvps = inv(transformationmatrix(plot)[]) * inv_f32_convert_matrix(plot, space) *
         clip_to_space(cam, space) * space_to_clip(cam, :pixel)
-    pre_transform = mvps * to_ndim(Vec4f, to_ndim(Vec3f, p, 0.0), 1.0)
-    p3 = Point3f(pre_transform) / pre_transform[4]
+    pre_transform = mvps * to_ndim(Vec4d, to_ndim(Vec3d, p, 0.0), 1.0)
+    p3 = Point3d(pre_transform) / pre_transform[4]
     return apply_transform(itf, p3, space)
 end
